Change expected behavior when deleting plugin
When a plugin is deleted, we previously expected that plugin to
disappear. That makes the delete plugin command have less utility than
it could.

Instead, only delete the plugin from the winning mod (in cases where
that plugin is provided by multiple mods).

This might be confusing for users when they have plugins provided by
multiple mods, but if they really want to nuke the plugin, they'll now
need to run the delete command for as many times as they have mods which
provide a plugin of that name.

+++ ammo/controller/bethesda.py
@@ -732,16 +732,16 @@ class BethesdaController(ModController):
         Removes specified plugin from the filesystem.
         """
 
-        def get_plugin_files(plugin):
+        def get_plugin_file(plugin) -> Path:
             """
-            Get plugin files from all enabled mods.
+            Get the plugin file from the winning mod.
             """
-            for mod in self.mods:
+            for mod in self.mods[-1:]:
                 if not mod.enabled:
                     continue
                 for file in mod.plugins:
                     if file.name == plugin.name:
-                        yield file
+                        return file
 
         if index == "all":
             deleted_plugins = ""
@@ -757,7 +757,7 @@ class BethesdaController(ModController):
                     self.do_refresh()
                     self.do_commit()
                 self.plugins.remove(plugin)
-                for file in get_plugin_files(plugin):
+                if (file := get_plugin_file(plugin)) is not None:
                     with ignored(FileNotFoundError):
                         log.info(f"Deleting PLUGIN: {file}")
                         file.unlink()
@@ -773,7 +773,7 @@ class BethesdaController(ModController):
                 raise Warning("You can only delete visible components.")
 
             self.plugins.remove(plugin)
-            for file in get_plugin_files(plugin):
+            if (file := get_plugin_file(plugin)) is not None:
                 with ignored(FileNotFoundError):
                     log.info(f"Deleting PLUGIN: {file}")
                     file.unlink()
